Update temporatory codes

+++ model/Shallow_Prompt_CLIP_PGD.py
@@ -385,7 +385,7 @@ class PromptCLIP_Shallow:
         for batch in tqdm(self.test_loader, desc="Evaluating CLIP Baseline with 'a photo of a {c}'", leave=False):
             images, labels = self.parse_batch(batch) 
             
-            image_features_baseline = self.model.encode_image(images).type(self.dtype)
+            image_features_baseline = self.model.encode_image(images).type(self.model.dtype)
             image_features_baseline = image_features_baseline / image_features_baseline.norm(dim=-1, keepdim=True)
 
             logits = logit_scale * image_features_baseline @ text_features_baseline.t()
